+++ lib/rive.dart
@@ -12,6 +12,7 @@ export 'package:rive/src/rive_core/animation/linear_animation_instance.dart';
 export 'package:rive/src/rive_core/animation/loop.dart';
 export 'package:rive/src/rive_core/animation/state_machine.dart';
 export 'package:rive/src/rive_core/artboard.dart';
+export 'package:rive/src/rive_core/assets/audio_asset.dart';
 export 'package:rive/src/rive_core/assets/font_asset.dart';
 export 'package:rive/src/rive_core/assets/image_asset.dart';
 export 'package:rive/src/rive_core/event.dart';